src/requirements.txt
Flask==1.1.2
google-auth==1.19.2
google-auth-oauthlib==0.4.1
gunicorn==20.0.4